Split the sql browser into code and results panels; with results panels going in the 'tools' notebook
this was done so that (1) the auto complete keyboard shortcuts works (2) files can be opened from the file system and then run. since files are opened by Notebook class
and NotebookClass alway opens them in a CodeControlClass the notebook class didn't know anything about the SqlBrowserPanel class; but we want to keep it that way.

AppFrame now propagates the AUI Notebook Changed event to all plugins; this was needed so that the SqlBrowser could change to the appropriate results tab based
on the current code control tab