Fixed a couple of problems with the linter
Removed 'not' it is not a PHP keyword
The heredocs were not being tokenized correctly if they did not end with a semicolon]
Fixed yet another bug in the UCharBufferedFile class; this time the buffer was not
being copied correctly when it was almost full when AppendToLexeme was called. The
buffer's slack was removed; but when the buffer grew the file read started from the
start of the new space; but it is supposed to overwrite the characters moved over because
of the removal of the slack space.