Added a logger for editor messages
Used the wxWidgets logging mechanism to log stuff to a new plugin.  The new logger is now
used in a couple of places where errors occurr, like when the PHP/explorer binaries cannot be invoked,
a new thread cannot be started, or a file cannot be opened.

Re-worked the file opening code so that a  code control does not get created if file opening fails.
Project / DB detection errors are noe propagated up from the Project class; the app class now logs
them using the new logging mechanism