Fixed a crash bug that appeared when a project was opened while a find in files was running
This was actually a much bigger issue having to do with the way threads were being handled. Needed to have a way to
emulate joinable threads, since we need to take care and make sure a thread stops running before continuing with
object destruction.

In this case, for example, the thread's delete method was called, but since this is a non-blocking operation the
FindInFilesPanelClass was cleaned up while the thread was still running, since it had not gotten to the TestDestroy()
check