Added versioning into the sqlite db files; bug fix during tag searching.
The tag finders where generating sql errors because of the table differences
between the tag cache db and the detector cache db.

versioning will work by checking a schema_version table and comparing
the row in the table with the hardcoded value in the source code.
Also, with this change the version check wil be done on program start,
that way the rest of the code does not have to perform this check (this
was the main reason for so many changes).