tweaks to project list dialog
select all button
remove multiple button
when adding multiple projects, the user can choose which directories to
add
project.
When the user clicks "OK" after enabling many projects, the editor
seemed to hang for a little bit. This is due to the way projects are
compared so that we can know which projects have just been edited.