properly re-tag files when renamed
when a file was renamed, it would never re-tag the file, but it needs
to under certain cases. For example, if file "a.php3" was renamed to "a.php"
we need to retag if ".php3" is not registered as a php extension.

Also, the tag parser is more tolerant of a file being deleted while it is being
tagged.