better handling of file renames that are not really renames
checking to see if this rename was an actual permanent rename as opposed to
file shuffling by apps.
some apps (notably VIM) will do unexpected things like rename a file when
they open it, then right away create a new file with the original name. if we captured a file rename
but the old file still exists, assume that the old file was put right back.