premake changes
the wx-config path can now be controlled via an environment variable (MVCEDITOR_WXCONFIG)
this is done so that buildbot can use the variable when doing the build; in the buildbot
slave we will compile wxWidgets once and then on every build we won't have too.

no longer run the unit tests as as post build action. there are too many, it takes too long,
plus we have buildbot that does it for us.