fix editor drawing issues on MSW
We now update the status bar info (line, column)
at specific intervals instead of in a EVT_STC_UPDATEUI event.
Updating the status bar text triggered a refresh of the whole
status bar (and it seems that the entire app is refreshed too)
and it makes the app feel sluggish.