Added a new user dialog.  It will popup when the user first runs the program.
The dialog will ask the user for a location to store settings; php
file extensions and PHP location.

The preferences panel was updated to have the settings from the new user
dialog (settings directory, php not installed).

There is a new option to disable running PHP scripts (PHP not installed)