updates to total search result dialog
in MSW, when the list control lost focus, the blue hightlight was
no longer shown.  this makes it difficult for the user to see
the selection when the user hits the down/up arrows to cycle
through the results.

ended up using a listbox instead of a list control, losing the
ability to show images for each tag