when a file has been created/deleted/renamed externally, update the modal explorer panel
fixed an assertion with handling external renames,
when a file was externally renamed, we captured the "modify"
event as well; the editor saw the file as both
modified and the code attempted. since the code handles
modifications by reparsing them, the new file was already
in the cache by the time the rename was handled. the fix
was to "collapse" modified events from files that were renamed