php doc comment panel enhancements
closing button works properly on linux.

fixed crash bug, having to do with  GTK sending log events
in a background.  retro-fitted the messages feature to
post log events instead of rendering them in the log handler.

added a link to the PHP.net site when the docs for a  native
function are shown