file mod checks feature bug fix
yet again ... this really bites...

broke up the watcher vs. the opened files checks
into their own features.  Now, fs watcher feature
only deals with getting FS watcher events and propagating
them, and the File mod feature only event deals with
files that are opened.

did this because every so often, the "file externally
modified" dialog popped up in MSW when the project was
in a local volume.

the only bad thing about this is that the user will only
get prompted for opened files when the app is reactivated
but looking at the source of other editors this is how
other editors do it; plus I really want to prevent these
bogus "file is extenally modified" prompts.