php code completion enhancement
there is a fallback algorithm for building the symbol table for a
file in the case where the file contains a syntax error. If there
is a syntax error, we use tokens to determine variables.  This
is not fool-proof, but it will allow php code completion
to show some suggestions when the file has a syntax error.