build sytem (packager) updates
packages now default to building packages of the current branch.
This makes it so that we can trigger building of packages
via buildbot.

Decoupled the version number from the tag/branch name.
The version number must follow certain guidelines depending
on which platform we are using, so we want the flexibility to
name the version differently than the tag/branch we are using. For example,
debian versions must be 'N.N.N' where N is a number [0-9], but we
may want to build a 'nightly' using the master branch, or we may
want to build a release candidate from the '0.6.2-RC' branch.